# ci: Speed up the CI by building tests without opt