--- a/.github/workflows/edr-ci.yml
+++ b/.github/workflows/edr-ci.yml
@@ -83,6 +83,14 @@ jobs:
             **/edr-cache
           key: edr-rs-rpc-cache-v1
 
+      - name: Build cargo tests
+        uses: actions-rs/cargo@v1
+        env:
+          RUSTFLAGS: ${{ matrix.RUSTFLAGS }}
+        with:
+          command: test
+          args: --workspace --all-targets --all-features --no-run
+
       - name: Run cargo test
         uses: actions-rs/cargo@v1
         env: