--- a/.github/workflows/test-recent-mainnet-block.yml
+++ b/.github/workflows/test-recent-mainnet-block.yml
@@ -4,9 +4,6 @@ on:
   schedule:
     - cron: "0 */8 * * *"
   workflow_dispatch:
-  push:
-    branches:
-      - chore/decrease-test-time
 
 defaults:
   run: