--- a/Cargo.toml
+++ b/Cargo.toml
@@ -10,5 +10,3 @@ rpath = true
 [profile.release]
 rpath = true
 
-[profile.test]
-opt-level = 0