--- a/crates/edr_provider/tests/issues/issue_356.rs
+++ b/crates/edr_provider/tests/issues/issue_356.rs
@@ -1,5 +1,3 @@
-#![cfg(all(feature = "test-remote", feature = "test-utils"))]
-
 use std::str::FromStr;
 
 use anyhow::Context;