--- a/crates/edr_provider/tests/issues/mod.rs
+++ b/crates/edr_provider/tests/issues/mod.rs
@@ -1,0 +1,10 @@
+#![cfg(all(feature = "test-remote", feature = "test-utils"))]
+
+mod issue_324;
+mod issue_325;
+mod issue_326;
+mod issue_346;
+mod issue_356;
+mod issue_361;
+mod issue_384;
+mod issue_407;