Closes #439 

This implements the idea outlined in https://github.com/NomicFoundation/edr/issues/439#issuecomment-2100679905, which sets `opt-level` accordingly in the `test` Cargo profile.

I also had an idea to use the sparse crates.io checkout but didn't realize that in addition to it being stabilized, it was also immediately set as a default in Rust 1.68, so we should be using that already now.

To sum up some of the timings for building/running the tests (unit + integration + doctests):

| OS \ opt-level         | [2 (warm; baseline)](https://github.com/NomicFoundation/edr/actions/runs/9416944157/job/25941201261) | [1](https://github.com/NomicFoundation/edr/actions/runs/9417516559/attempts/1)       | [1 (second run)](https://github.com/NomicFoundation/edr/actions/runs/9417516559/attempts/2) | [0](https://github.com/NomicFoundation/edr/actions/runs/9418304660/attempts/1)      | [0 (second run)](https://github.com/NomicFoundation/edr/actions/runs/9418304660/attempts/2) |
|------------------------|--------------------|---------|------------|--------|------------|
| ubuntu-latest (build)  | 10m 8s             | 10m 50s | 10m 43s    | 4m 14s | 4m 8s  **(-59%)**     |
| ubuntu-latest (run)    | 47s                | 43s     | 44s        | 44s    | 43s        |
| macos-latest (build)   | 18m 32s            | 8m 36s  | 8m 38s     | 2m 47s | 2m 25s **(-87%)**    |
| macos-latest (run)     | 1m 56s             | 44s     | 44s        | 44s    | 36s        |
| windows-latest (build) | 21m 4s             | 19m 29s | 18m 40s    | 7m 53s | 6m 16s   **(-70%)**   |
| windows-latest (run)   | 1m 25s             | 1m 22s  | 1m 21s     | 49s    | 48s        |
| _total workflow time_   | _26m 44s_             | _25m 26s_  | _23m 39s_     | _13m 28s_    | _10m 56s_        |


1. What's surprising is the macos being overly slow at first on baseline and faster than ubuntu later - if I had to guess that may be some noise from other tenants?
2. What's also surprising is that time spent running also decreases when we decrease the opt-level. One guess would be that we keep using the default of `codegen-units = 256` but not run LTO and we end up bloating the binary, which is less cache friendly? That shouldn't have that much impact, though...

Despite no. 2 which should probably be investigated a little bit more at least locally, I think the build time savings are clear and much more outweigh the time spent running the tests themselves. Unless we're *very* off with test runtime, the rate of improvement should also translate to time savings when ran locally as well.

cc @Wodann since you originally proposed dropping the `opt-level` for `profile.test`