@Xanewok 2024-06-07
@fvictorio  I've marked this as closing #439 since we're around the 10 min mark, which seems fine now. If we still want to pursue further improvements, feel free to edit the OP to drop the `Closes` remark.