@Wodann 2024-06-10
1. Yes, that's where the cacheable RPC requests are cached locally.
2. That sounds like a good idea. Feel free to combine all of the issue test files into one :)