@Xanewok 2024-06-10
Combined some of the issues tests and got some good savings - there's still room for improvement but I'd consider this to be more or less done until we want to push it to extreme. For now, we got 9 minute run with Windows test taking the most (> 6 minutes), so the CI improvements seems sufficient for now.